requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
numpy>=1.24.3
opencv-python>=4.7.0
pillow>=9.5.0
tensorflow-cpu>=2.12.0
fastapi[standard]